Add files via upload
The global tourism industry has experienced a lot of concomitant serge in the waste generation in the period of past few years. And this is not in any specific region it has come up in front of the industry as a challenge that has to be mitigated on a global scale along with raising the concerns related to accommodations, food services, conference and entertainment services. The key products of the waste generated include food, cooking oil, glass, cardboard, paper, aluminum, plastic, and dangerous liquids including lamp oil and cleaning chemicals. The amount of garbage generated by the hotel industry is expanding rapidly due to increased worldwide travel. For instance, approximately a million tons of garbage are produced year by hotels in the United States. The report had stated that, with the identified hospitality industries to keep up with the guidelines mentioned, it had been responsible to showcase that they were responsible ad ethical for the task. Thus, the report had even determined how these rules were crucial for creating a united and effective approach for reducing extreme amount of waste being generated by any of the hospitality industry across the world. For considerably fixing these environmental issues, the report here had even provided suitable recommendation for creating a strong and sustainable procedures that the respective hospitality sectors could utilize for managing their waste generation, thereby following a circular economy idea.